refactor(session): split cleanup.ts by lifecycle phase
`src/session/cleanup.ts` was 523 lines with five overlapping functions
whose names did not encode WHEN they run or WHICH files they may touch
(`cleanupStaleSession`, `cleanupSession`, `cleanupStaleDaemonPid`,
`forceRecoverStaleSession`, `performSessionCleanup`). Readers had to
grep doc comments to pick the right one.

Split by phase with names that say when each runs:

  cleanup/preStart.ts      — before daemon launch (file-lock only; may
    remove shared daemon files if dead).
    cleanupStaleSession      → cleanupBeforeDaemonStart
    cleanupStaleDaemonPid    → cleanupStaleDaemonArtifacts
  cleanup/runtime.ts       — while daemon is alive (session files only;
    daemon files MUST survive).
    forceRecoverStaleSession → recoverStaleSessionAtRuntime
  cleanup/postSession.ts   — after clean session end (all session +
    daemon files removable).
    cleanupSession           → cleanupAfterSessionEnd
  cleanup/userCommands.ts  — orchestrator for `bdg stop` / `bdg cleanup`
    (composes the phases based on flags).
    performSessionCleanup    (unchanged name)
  cleanup/primitives.ts    — shared SIGKILL + chrome-pid helpers at the
    bottom of the dependency graph.

No back-compat barrel (CLAUDE.md forbids); all 7 call sites updated in
the same commit. `ISessionService.forceRecoverStaleSession` keeps its
method name as the handler-facing API; the underlying implementation is
the renamed cleanup function.
