+++ src/commands/status.ts
@@ -7,7 +7,7 @@ import { isDaemonConnectionError } from '@/errors/index.js';
 import { invalidResponseError, daemonNotRunningError } from '@/errors/messages.js';
 import { getStatus } from '@/ipc/client.js';
 import type { SessionActivity, PageState } from '@/ipc/index.js';
-import { cleanupStaleDaemonPid } from '@/session/cleanup.js';
+import { cleanupStaleDaemonArtifacts } from '@/session/cleanup/preStart.js';
 import type { SessionMetadata } from '@/session/metadata.js';
 import {
   formatSessionStatus,
@@ -104,7 +104,7 @@ export function registerStatusCommand(program: Command): void {
           } catch (error) {
             const errorMessage = getErrorMessage(error);
             if (isDaemonConnectionError(error)) {
-              const cleaned = cleanupStaleDaemonPid();
+              const cleaned = cleanupStaleDaemonArtifacts();
               return {
                 success: false,
                 error: daemonNotRunningError({ staleCleanedUp: cleaned }),
